requirements.txt
# Production Requirements
Django>=5.0.6,<5.0.9
djangorestframework>=3.15.0,<3.15.2
pymongo>=4.5.0,<4.6.0
mongoengine>=0.27.0,<0.28.0